Introduction - Simulation

+++ 1.algorithm/3.introduction/1.simulation/A1009.cpp
@@ -23,27 +23,27 @@ using namespace std;
 
 int main(int argc, char const *argv[]) {
 
-    map<int, float> a;
+    map<int, double> a;
     int k1;
     cin >> k1;
     for (int i = 0; i < k1; i++) {
         int a1;
-        float a2;
+        double a2;
         cin >> a1 >> a2;
         a[a1] = a2;
     }
-    map<int, float, greater<int>> b;
+    map<int, double, greater<int>> b;
     int k2;
     cin >> k2;
     for (int i = 0; i < k2; i++) {
         int a1;
-        float a2;
+        double a2;
         cin >> a1 >> a2;
         for (auto it : a) {
             b[a1 + it.first] += a2 * it.second;
         }
     }
-    int cnt = count_if(b.begin(), b.end(), [](pair<int, float> b) -> bool {
+    int cnt = count_if(b.begin(), b.end(), [](pair<int, double> b) -> bool {
         return b.second != 0;
     });
     cout << cnt;
